// brokerchain-1.0.0/src/library/cryptography/ecdh-x25519-secret-compute/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {
    my_private_key_hex: string;
    target_public_key_hex: string;
}

export interface OutputOk {
    secret: { hex: string; base64: string };
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            if (field === "my_private_key_hex") return;
            if (field === "target_public_key_hex") return;
            throw new Error("v contains unknown field: " + field);
        });

        log.println("v.my_private_key_hex must be string");
        if (typeof v.my_private_key_hex !== "string") {
            throw new Error("v.my_private_key_hex is not string");
        }

        log.println("v.target_public_key_hex must be string");
        if (typeof v.target_public_key_hex !== "string") {
            throw new Error("v.target_public_key_hex is not string");
        }
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {
            secret: copy_secret(v.secret)
        };
        return obj;
    } else {
        return { secret: { hex: "", base64: "" } };
    }

    function copy_secret(v: any): { hex: string; base64: string } {
        if (typeof v === "object" && v !== null) {
            const obj = {
                hex: copy_hex(v.hex),
                base64: copy_base64(v.base64)
            };
            return obj;
        } else {
            return { hex: "", base64: "" };
        }

        function copy_hex(v: any): string {
            return typeof v === "string" ? v : "";
        }

        function copy_base64(v: any): string {
            return typeof v === "string" ? v : "";
        }
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;